Fixed problem with merging suffixes: suffix merging is now triggered after the entire tree is built.
For that to work, we keep a list of all leaf items of the produced NFA and then merge these leaves together. Whenever merging two leaves, we recursively attempt to merge successor states.
When we merge suffixes while building the tree, we potentially produce wrong output as we might extend joint paths with output that is not intended for all merged paths.