Cutting setup time
Was especially high in A2 (100s), now down to 3s, with only minor performance losses in the other stages.

Problem was that we iterated over every PathState in the setup. With a max of 330k states in A2, doing that for every pivot element of every sequence is not a good idea.