Changed DesqDfs to build FST/DFA only once per JVM
This is a first version. It will fail when there are multiple pattern expressions (and therefore, multiple different FSTs/DFAs) within one application, as FSTs and DFAs won't be reconstructed for the 2nd and following pattern expressions. 
As long as the pattern expressions run in separate applications, everything should work fine, as executor JVMs are not shared between applications.